+++ lib/data layer/services/settings_service.dart
@@ -1,4 +1,4 @@
-import 'package:donotimagine/data%20layer/models/theme_settings_model.dart';
+import 'package:donotimagine/data%20layer/models/settings/theme_settings_model.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -11,13 +11,13 @@ mixin ThemeSettingsMixin {
       'theme_setting_dark',
       ThemeSettingsModel().themeMode == ThemeMode.dark,
     );
-    await prefs.setInt('theme_setting_dark', ThemeSettingsModel().fontSize);
+    await prefs.setDouble('theme_setting_dark', ThemeSettingsModel().fontSize);
   }
 
   Future<void> loadThemeSettings() async {
     final prefs = await SharedPreferences.getInstance();
     final isDarkTheme = prefs.getBool('theme_setting_dark') ?? false;
-    final fontSize = prefs.getInt('theme_setting_dark') ?? 16;
+    final fontSize = prefs.getDouble('theme_setting_dark') ?? 16;
     ThemeSettingsModel().set(
       fontSize: fontSize,
       themeMode: isDarkTheme ? ThemeMode.dark : ThemeMode.light,
